From now on only the latest version can be installed.
The name of the package now does not include the version number any more, since it is not expected that one wants to install an old watchdog server version. Furthermore, the maintenace scripts are updated.
The main change is that the service file used by systemd is moved to the configuration part of the package. This allows to change it permanetly by the user (e.g. require rpcbind for watchdog operation).
Finally the persist file is stored event if the package is removed. After reinstallation it will be considered again.