Corrected dependencies-dev for some doocs-depending packages
If and only if linking against the DOOCS libraries is required and the libraries are specified in the ${PROJECT_NAME}_LIBRARIES cmake variable the dev package needs to depend on those libraries. In two packages libgul14 was missing (which is not correctly pulled in by the serverlib), in a third package the libraries were specified redundantly.