fix(command-line-tools): insufficient dependency spec
The mtca4u-command-line-tools package contains the symlink to the
executable in the mtca4u-command-line-tools<version> package. The
version requirement for the dependency between the two packages must be
exact, since otherwise updating just the mtca4u-command-line-tools
will not trigger an update of the mtca4u-command-line-tools<version>
resulting in a dangling symlink.

This commit also increases artificially the build number to allow
re-publication without source code change.